pyproject.toml
[project]
name = "expenselab"
version = "0.1.0"
description = ""
authors = [
    {name = "katsiarynach",email = "katsiaryna.chubarava@innowise.com"}
]
requires-python = ">=3.10"
dependencies = [
    "streamlit (>=1.53.1,<2.0.0)",
    "pandas-stubs (>=2.3.3,<2.4.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
